+++ src/InicioSesion.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="inicioSesion">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="InicioSesion">
   <grid id="27dc6" binding="JPanelprincipalInicioSesion" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -56,17 +56,6 @@
                   <text value="USUARIO:"/>
                 </properties>
               </component>
-              <component id="f38e5" class="javax.swing.JTextField" binding="textField1" default-binding="true">
-                <constraints>
-                  <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-                    <preferred-size width="150" height="-1"/>
-                  </grid>
-                </constraints>
-                <properties>
-                  <background color="-395009"/>
-                  <text value=""/>
-                </properties>
-              </component>
               <component id="ceadb" class="javax.swing.JLabel" binding="logoMarcostv">
                 <constraints>
                   <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="2" fill="0" indent="0" use-parent-layout="false"/>
@@ -84,6 +73,16 @@
                 <properties>
                   <background color="-395009"/>
                   <text value="Iniciar sesión"/>
+                </properties>
+              </component>
+              <component id="60b06" class="javax.swing.JPasswordField" binding="introContrasenya">
+                <constraints>
+                  <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
+                    <preferred-size width="150" height="-1"/>
+                  </grid>
+                </constraints>
+                <properties>
+                  <background color="-395009"/>
                 </properties>
               </component>
             </children>
